requirements_txt.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
youtube-transcript-api==0.6.1
openai==0.28.1
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0
